requirements.txt
asgiref==3.3.4
attrs==19.3.0
Automat==0.8.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
discord-logger==1.2.3
discord-webhook==0.13.0
distro==1.4.0
distro-info==1.0
Django==3.2.4
django-crispy-forms==1.12.0
django-ipware==4.0.2
django-mathfilters==1.0.0
django-nested-inline==0.4.3
dnspython==2.1.0
entrypoints==0.3
et-xmlfile==1.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
jeepney==0.7.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
more-itertools==4.2.0
netaddr==0.8.0
netifaces==0.10.4
numpy==1.20.3
oauthlib==3.1.0
openpyxl==3.0.7
pandas==1.2.4
pbr==5.8.1
psutil==5.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pydbus==0.6.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-dateutil==2.8.1
python-gnupg==0.4.8
python-xlib==0.29
pythondialog==3.5.1
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.1
ssh-import-id==5.10
testresources==2.0.1
Twisted==18.9.0
urllib3==1.25.8
wadllib==1.3.3
xdg==5
zipp==1.0.0
zope.interface==4.7.1
